# backend/openapi_cache.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "链客宝API",
    "description": "链客宝后端API服务 — Premium Business Network and Entrepreneur Supply-Demand Matching Platform。\n\n提供认证、产品、订单、搜索、支付、充值、发票、对账、联系人管理、活动时间线、数据洞察、供需匹配、AI智能匹配引擎、推广员体系、管理后台、系统配置等完整业务能力。",
    "contact": {
      "name": "链客宝团队",
      "url": "https://www.go-aiport.com/",
      "email": "support@go-aiport.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://www.go-aiport.com/"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/payment/wxpay/unified-order": {
      "post": {
        "tags": [
          "支付"
        ],
        "summary": "Wxpay Unified Order",
        "description": "微信统一下单 (JSAPI)\n\n使用 IJPay WxPayApi V3 接口。\n若配置不完整或调用失败，降级为 mock 模式。",
        "operationId": "wxpay_unified_order_api_payment_wxpay_unified_order_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WxPayUnifiedOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/payment/wxpay/query/{order_no}": {
      "get": {
        "tags": [
          "支付"
        ],
        "summary": "Wxpay Query",
        "description": "查询订单支付状态",
        "operationId": "wxpay_query_api_payment_wxpay_query__order_no__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/payment/wxpay/refund": {
      "post": {
        "tags": [
          "支付"
        ],
        "summary": "Wxpay Refund",
        "description": "微信退款",
        "operationId": "wxpay_refund_api_payment_wxpay_refund_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WxPayRefundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/payment/alipay/unified-order": {
      "post": {
        "tags": [
          "支付"
        ],
        "summary": "Alipay Unified Order",
        "description": "支付宝统一下单 (APP 支付)",
        "operationId": "alipay_unified_order_api_payment_alipay_unified_order_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AliPayUnifiedOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AliPayUnifiedOrderRequest": {
        "properties": {
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          }
        },
        "type": "object",
        "required": [
          "order_id"
        ],
        "title": "AliPayUnifiedOrderRequest",
        "description": "支付宝统一下单请求"
      },
      "ApiResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse",
        "description": "统一API响应格式"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WxPayRefundRequest": {
        "properties": {
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "order_id"
        ],
        "title": "WxPayRefundRequest",
        "description": "微信退款请求"
      },
      "WxPayUnifiedOrderRequest": {
        "properties": {
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "openid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openid"
          }
        },
        "type": "object",
        "required": [
          "order_id"
        ],
        "title": "WxPayUnifiedOrderRequest",
        "description": "微信统一下单请求"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "https://www.go-aiport.com",
      "description": "生产环境"
    },
    {
      "url": "https://staging.go-aiport.com",
      "description": "预发布环境"
    },
    {
      "url": "http://localhost:7800",
      "description": "本地开发环境"
    }
  ]
}